tests/wpt/web-platform-tests/shadow-dom/untriaged/elements-and-dom-objects/shadowroot-object/shadowroot-methods/test-001.html

<!DOCTYPE html>
<!--
Distributed under both the W3C Test Suite License [1] and the W3C
3-clause BSD License [2]. To contribute to a W3C Test Suite, see the
policies and contribution forms [3].

[1] http://www.w3.org/Consortium/Legal/2008/04-testsuite-license
[2] http://www.w3.org/Consortium/Legal/2008/03-bsd-license
[3] http://www.w3.org/2004/10/27-testcases
-->
<html>
<head>
<title>Shadow DOM Test: A_10_01_02_01</title>
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="help" href="http://www.w3.org/TR/2013/WD-shadow-dom-20130514/#shadow-root-methods">
<meta name="assert" content="ShadowRoot Object: HTMLElement getElementById(DOMString elementId) method">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
test(function () {

	var d = newHTMLDocument();

    var el = d.createElement('div');
    d.body.appendChild(el);

    var s = el.createShadowRoot();

    var child = d.createElement('span');
    child.setAttribute('id', 'span_id');
    s.appendChild(child);

    assert_true(s.getElementById('span_id') != null, 'Point 1: ShadowRoot getElementById() ' +
    		'method should return child element');
    assert_equals(s.getElementById('span_id').getAttribute('id'), 'span_id', 'Point 2: ' +
    		'ShadowRoot getElementById() method should return child element');

}, 'A_10_01_02_01_T01');



test(function () {

	var d = newHTMLDocument();

    var el = d.createElement('div');
    d.body.appendChild(el);

    var s = el.createShadowRoot();

    assert_true(s.getElementById('span_id') == null, ' ShadowRoot getElementById() ' +
    		'method should return null if matching element not found');

}, 'A_10_01_02_01_T02');
</script>
</body>
</html>
